# common-platform-rule-engine-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-rule-engine
  description: Rule Engine Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/rules
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/validation/yaml:
    post:
      tags:
        - YAML DSL Validation
      summary: Validate YAML DSL Rule
      description: |
        Performs comprehensive static analysis and validation of a YAML DSL rule.
        Acts as a static code analyzer to catch syntax errors, naming convention violations,
        dependency issues, logic problems, and provide optimization suggestions.

        This endpoint helps developers:
        - Catch errors before deployment
        - Ensure naming convention compliance
        - Detect order-of-operations issues (like the bug we just fixed!)
        - Optimize rule performance
        - Follow best practices
      operationId: validateYaml
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        description: YAML DSL content and validation options
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateYamlRequest'
            examples:
              Basic Credit Scoring Rule:
                description: A simple credit scoring rule with proper naming conventions
                value:
                  yamlContent: |-
                    name: Credit Score Assessment
                    description: Evaluate credit worthiness based on score and income
                    version: 1.0.0
                    inputs:
                      - creditScore
                      - annualIncome
                      - employmentYears
                      - existingDebt
                    when:
                      - creditScore at_least MIN_CREDIT_SCORE
                      - annualIncome at_least 50000
                      - employmentYears at_least 2
                    then:
                      - calculate debt_to_income as existingDebt / annualIncome
                      - set is_eligible to true
                      - set approval_tier to "STANDARD"
                    else:
                      - set is_eligible to false
                      - set approval_tier to "DECLINED"
                    output:
                      is_eligible: is_eligible
                      approval_tier: approval_tier
                      debt_to_income: debt_to_income
                  includeSuggestions: true
                  includeMetrics: true
              Rule with Issues:
                description: A rule with various validation issues to demonstrate error detection
                value:
                  yamlContent: |-
                    name: Bad Rule with Naming Issues
                    inputs:
                      - CREDIT_SCORE
                      - annual_income
                      - employment_years
                    when:
                      - CREDIT_SCORE at_least 650
                      - debt_to_income less_than 0.4
                    then:
                      - calculate DEBT_RATIO as monthlyDebt / annual_income
                      - set FINAL_DECISION to "APPROVED"
                  categories:
                    - SYNTAX
                    - NAMING
                    - DEPENDENCIES
                  minSeverity: WARNING
        required: true
      responses:
        '200':
          description: Validation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '400':
          description: Invalid request (empty YAML, too large, etc.)
          content:
            application/json: {}
        '500':
          description: Internal server error during validation
          content:
            application/json: {}
...
  /api/v1/validation/syntax:
    get:
      tags:
        - YAML DSL Validation
      summary: Quick YAML Syntax Check
      description: |
        Performs a quick syntax-only validation of YAML DSL content.
        Useful for real-time validation in editors or IDEs.
        Only checks basic YAML syntax and DSL structure, not business logic.
      operationId: checkSyntax
      parameters:
        - name: yaml
          in: query
          description: YAML DSL content to validate (URL encoded)
          required: true
          schema:
            maxLength: 50000
            minLength: 0
            type: string
          example: 'name: Test Rule\ninputs: [creditScore]\nwhen: [creditScore at_least 650]\nthen: [set result to APPROVED]'
      responses:
        '200':
          description: Syntax check completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '400':
          description: Invalid request parameters
          content:
            application/json: {}
...
components:
  schemas:
...
    ValidateYamlRequest:
      required:
        - yamlContent
      type: object
      properties:
        yamlContent:
          maxLength: 100000
          minLength: 0
          type: string
        categories:
          uniqueItems: true
          type: array
          items:
            type: string
            enum:
              - SYNTAX
              - NAMING
              - DEPENDENCIES
              - LOGIC
              - PERFORMANCE
              - BEST_PRACTICES
        minSeverity:
          type: string
          enum:
            - CRITICAL
            - ERROR
            - WARNING
            - INFO
        includeSuggestions:
          type: boolean
        includeMetrics:
          type: boolean
    ValidationIssue:
      type: object
      properties:
        code:
          type: string
        severity:
          type: string
          enum:
            - CRITICAL
            - ERROR
            - WARNING
            - INFO
        message:
          type: string
        description:
          type: string
        location:
          $ref: '#/components/schemas/ValidationLocation'
        suggestion:
          type: string
        examples:
          type: array
          items:
            type: string
    ValidationIssues:
      type: object
      properties:
        syntax:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        naming:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        logic:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        performance:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        bestPractices:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
    ValidationLocation:
      type: object
      properties:
        section:
          type: string
        lineNumber:
          type: integer
          format: int32
        path:
          type: string
        context:
          type: string
    ValidationMetadata:
      type: object
      properties:
        validatorVersion:
          type: string
        validatedAt:
          type: string
          format: date-time
        validationTimeMs:
          type: integer
          format: int64
        ruleName:
          type: string
        statistics:
          type: object
          additionalProperties:
            type: object
    ValidationResult:
      type: object
      properties:
        status:
          type: string
          enum:
            - VALID
            - WARNING
            - ERROR
            - CRITICAL_ERROR
        summary:
          $ref: '#/components/schemas/ValidationSummary'
        issues:
          $ref: '#/components/schemas/ValidationIssues'
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/ValidationSuggestion'
        metadata:
          $ref: '#/components/schemas/ValidationMetadata'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationIssue'
        valid:
          type: boolean
        errorSummary:
          type: string
    ValidationSuggestion:
      type: object
      properties:
        category:
          type: string
        title:
          type: string
        description:
          type: string
        example:
          type: string
        impact:
          type: integer
          format: int32
    ValidationSummary:
      type: object
      properties:
        totalIssues:
          type: integer
          format: int32
        criticalErrors:
          type: integer
          format: int32
        errors:
          type: integer
          format: int32
        warnings:
          type: integer
          format: int32
        suggestions:
          type: integer
          format: int32
        qualityScore:
          type: number
          format: double
...